IT WAS THE RASTERIZER
Turns out rendering each cone to take the entirety of the screen makes a lot of work for the rasterizer. I've shrunk the radius of the cones, but there's a chance there will be holes in the image. Will have to adjust the radius dynamically based on the number of points.